name change and removing duplicate tests
