+++ test/registered/8-gpu-models/test_deepseek_v31.py
@@ -8,14 +8,16 @@ import sys
 import unittest
 from pathlib import Path
 
-sys.path.insert(0, str(Path(__file__).parent.parent))
+# Add nightly directory to path for run_combined_tests import
+sys.path.insert(0, str(Path(__file__).parent.parent.parent / "nightly"))
 
-from nightly_metrics import run_metrics
+from run_combined_tests import run_metrics
 
 from sglang.test.ci.ci_register import register_cuda_ci
 from sglang.test.test_utils import DEFAULT_URL_FOR_TEST, ModelLaunchSettings
 
-register_cuda_ci(est_time=12000, suite="nightly-8-gpu-temp-b200", nightly=True)
+# Registered to nightly-8-gpu-b200-basic suite (B200 only)
+register_cuda_ci(est_time=12000, suite="nightly-8-gpu-b200-basic", nightly=True)
 
 DEEPSEEK_V31_MODEL_PATH = "deepseek-ai/DeepSeek-V3.1"
