+++ test/registered/8-gpu-models/test_minimax_m2.py
@@ -9,16 +9,17 @@ import sys
 import unittest
 from pathlib import Path
 
-sys.path.insert(0, str(Path(__file__).parent.parent))
+# Add nightly directory to path for run_combined_tests import
+sys.path.insert(0, str(Path(__file__).parent.parent.parent / "nightly"))
 
-from nightly_metrics import run_metrics
+from run_combined_tests import run_metrics
 
 from sglang.test.ci.ci_register import register_cuda_ci
 from sglang.test.test_utils import DEFAULT_URL_FOR_TEST, ModelLaunchSettings
 
-# Registered to nightly-8-gpu-temp suite for testing
+# Registered to nightly-8-gpu-h200-basic suite
 # This suite should be run with --timeout-per-file=12000 (200 minutes)
-register_cuda_ci(est_time=12000, suite="nightly-8-gpu-temp", nightly=True)
+register_cuda_ci(est_time=12000, suite="nightly-8-gpu-h200-basic", nightly=True)
 
 MINIMAX_M2_MODEL_PATH = "MiniMaxAI/MiniMax-M2"
