+++ test/registered/8-gpu-models/test_mistral_large3.py
@@ -10,16 +10,17 @@ import sys
 import unittest
 from pathlib import Path
 
-sys.path.insert(0, str(Path(__file__).parent.parent))
+# Add nightly directory to path for run_combined_tests import
+sys.path.insert(0, str(Path(__file__).parent.parent.parent / "nightly"))
 
-from nightly_metrics import run_metrics
+from run_combined_tests import run_metrics
 
 from sglang.test.ci.ci_register import register_cuda_ci
 from sglang.test.test_utils import DEFAULT_URL_FOR_TEST, ModelLaunchSettings
 
-# Registered to nightly-8-gpu-temp-b200 suite (B200 only - requires trtllm_mla backend)
+# Registered to nightly-8-gpu-b200-basic suite (B200 only - requires trtllm_mla backend)
 # This suite should be run with --timeout-per-file=12000 (200 minutes)
-register_cuda_ci(est_time=12000, suite="nightly-8-gpu-temp-b200", nightly=True)
+register_cuda_ci(est_time=12000, suite="nightly-8-gpu-b200-basic", nightly=True)
 
 MISTRAL_LARGE3_MODEL_PATH = "mistralai/Mistral-Large-3-675B-Instruct-2512"
 MISTRAL_LARGE3_EAGLE_MODEL_PATH = "mistralai/Mistral-Large-3-675B-Instruct-2512-Eagle"
