+++ test/run_suite.py
@@ -34,8 +34,8 @@ NIGHTLY_SUITES = {
         "nightly-8-gpu-h200",
         "nightly-8-gpu-h20",
         "nightly-8-gpu-b200",
-        "nightly-8-gpu-temp",  # Temporary suite for testing unified perf+accuracy tests (H200)
-        "nightly-8-gpu-temp-b200",  # Temporary suite for B200-only unified tests
+        "nightly-8-gpu-h200-basic",  # Basic perf+accuracy tests for large models on H200
+        "nightly-8-gpu-b200-basic",  # Basic perf+accuracy tests for large models on B200
     ],
     HWBackend.AMD: ["nightly-amd"],
     HWBackend.CPU: [],
